requirements.txt
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
orjson==3.8.6
platformdirs==3.0.0
pycodestyle==2.10.0
pydantic==1.10.5
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
xmltodict==0.13.0
zeep==4.2.1
